Moved feature to hologram provider, proxy method
