+++ api/src/main/java/de/minebench/tresor/services/hologram/Hologram.java
@@ -1,10 +1,12 @@
 package de.minebench.tresor.services.hologram;
 
+import de.minebench.tresor.services.hologram.Holograms.Feature;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
 
 public interface Hologram {
 
+    Holograms getProvider();
     String getHologramId();
 
     void addLine(String line);
@@ -23,10 +25,8 @@ public interface Hologram {
 
     void destroy();
 
-    boolean supports(Feature feature);
-
-    enum Feature {
-        PER_PLAYER,
+    default boolean supports(Feature feature) {
+        return getProvider().supports(feature);
     }
 
 }
